Added the Project.
This project contains implementation of a physics engine and a short game that makes use of it. State based AI that makes use of A* pathifinding are also implemented in the second level. Collider debugging gizmos were designed and can be visualised via the X key.
Via the F key, the quadtree used for broadphase optimisation can be viewed also